app/src/main/java/com/example/nytimes/data/network/api/NetworkHelper.kt
package com.example.nytimes.data.network.api

import com.squareup.moshi.Moshi
import kotlinx.coroutines.CoroutineDispatcher
import kotlinx.coroutines.withContext
import okio.IOException
import retrofit2.HttpException

abstract class NetworkHelper<in P, out R>(private  val coroutineDispatcher: CoroutineDispatcher) {

     suspend fun  execute( parameters: P? ): ResultWrapper<R> {
        return withContext(coroutineDispatcher) {
            try {
               buildUseCase(parameters)
            } catch (throwable: Throwable) {
                when (throwable) {
                    is IOException -> ResultWrapper.NetworkError
                    is HttpException -> {
                        val code = throwable.code()
                        val errorResponse = convertErrorBody(throwable)
                        ResultWrapper.GenericError(code, errorResponse)
                    }
                    else -> {
                        ResultWrapper.GenericError(null, null)
                    }
                }
            }
        }
    }

    abstract suspend fun buildUseCase(parameters : P?) : ResultWrapper<R>

    private fun convertErrorBody(throwable: HttpException): ErrorResponse? {
        return try {
            throwable.response()?.errorBody()?.source()?.let {
                val moshiAdapter = Moshi.Builder().build().adapter(ErrorResponse::class.java)
                moshiAdapter.fromJson(it)
            }
        } catch (exception: Exception) {
            null
        }
    }


}